+++ hfbm/hfbm.vcxproj
@@ -86,11 +86,21 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <QtRcc Include="gui.qrc" />
-    <QtUic Include="gui.ui" />
-    <QtMoc Include="gui.h" />
-    <ClCompile Include="gui.cpp" />
+    <ClCompile Include="GUI\gui.cpp" />
+    <ClCompile Include="Image2D\ImageViewer.cpp" />
     <ClCompile Include="main.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <QtMoc Include="GUI\gui.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <QtUic Include="GUI\gui.ui" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Image2D\ImageViewer.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <QtRcc Include="GUI\gui.qrc" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Condition="Exists('$(QtMsBuild)\qt.targets')">
